requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
pydantic>=2.0
python-dotenv>=1.0

langchain>=0.3
langchain-openai>=0.2
langchain-community>=0.2
langchain-classic>=0.0.30
langchain-text-splitters>=0.3

chromadb>=0.5
tiktoken>=0.7
rank-bm25>=0.2
langsmith>=0.1
